Sync android master branch code
1. Orignal commit message title:
	fixing a typo
2. commit hash id:
	efb347c9620f8b8f58e53771406a34327fd8c8e6
3. Original branch name:
	a. origin/master,
	b. origin/HEAD

+++ chromebase/base/json/json_parser.h
@@ -129,14 +129,14 @@ class BASE_EXPORT JSONParser {
 
   // Returns the next |count| bytes of the input stream, or nullopt if fewer
   // than |count| bytes remain.
-  Optional<StringPiece> PeekChars(size_t count);
+  Optional<StringPiece> PeekChars(int count);
 
   // Calls PeekChars() with a |count| of 1.
   Optional<char> PeekChar();
 
   // Returns the next |count| bytes of the input stream, or nullopt if fewer
   // than |count| bytes remain, and advances the parser position by |count|.
-  Optional<StringPiece> ConsumeChars(size_t count);
+  Optional<StringPiece> ConsumeChars(int count);
 
   // Calls ConsumeChars() with a |count| of 1.
   Optional<char> ConsumeChar();
